Содержание экрана description вынесено в отдельную View
